refactor: extract handler functions into separate file
